requirements.txt
cherrypy==18.1.1
PyYAML==6.0
psutil==5.4.8
numpy==1.21.6
Pillow==7.1.2
scipy==1.4.1
requests==2.23.0
tqdm==4.64.0
omegaconf==2.1.2
scikit-image==0.16.2
transformers==4.18.0
einops==0.4.1
pilgram==1.2.1
catboost==1.0.5
torch==1.10.0
torchvision==0.11.1
opencv-python==4.1.2.30
albumentations==0.1.12
tensorflow==1.15.0
keras==2.0.8
h5py==2.10.0
scikit-learn==1.0.2
